+++ clitest/ibc.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-echo "ibc test not implemented"